requirements.txt
fastapi
uvicorn
torch==2.0.0+cpu
torchvision==0.15.0+cpu
Pillow
python-multipart
numpy
mangum
boto3